fix indenting and separate out styles from component #57, #65
