+++ src/screens/Register.js
@@ -5,16 +5,16 @@ import {resetPassword} from './ResetPassword'
 import styles from '../styles/authStyles'
 
 export default function Register({navigation}) {
-	 const [email, setEmail] = useState("");
-	 const [password, setPassword] = useState("");
-	 const [confirmPassword, setConfirmPassword] = useState("")
-	 const [error, setError] =  useState("");
+	const [email, setEmail] = useState("");
+	const [password, setPassword] = useState("");
+	const [confirmPassword, setConfirmPassword] = useState("")
+	const [error, setError] =  useState("");
 
-	 useLayoutEffect(() => {
-		 navigation.setOptions({
-			 headerBackTitle: "Back to Login",
-		 })
-	 }, [navigation])
+	useLayoutEffect(() => {
+		navigation.setOptions({
+			headerBackTitle: "Back to Login",
+		})
+	}, [navigation])
 
 	const onSignUp = () => {
 		if (!email) {
@@ -25,51 +25,51 @@ export default function Register({navigation}) {
 			setError("Password and confirm password does not match. Try it again.")
 			return;
 		}
-		 firebase
-		 	.auth()
+		firebase
+			.auth()
 			.createUserWithEmailAndPassword(email, password)
-		 	.then((response) => {
-				 console.log("SignUp RESULT: ", response);
-				 const uid = response.user.uid
-				 const data = {
-					 email,
-					 cashOnHand: 0,
-					 positions: [],
-					 transactions: [],
-					 watchlist: [],
-				 };
-				 const usersRef = firebase.firestore().collection('users');
-				 usersRef
-				 	.doc(uid)
-					.set(data)
-					.then(() => {
-						navigation.navigate("Login")
-					})					
-				})
-			 .catch((error) => setError(error.message))
+			.then((response) => {
+				console.log("SignUp RESULT: ", response);
+				const uid = response.user.uid
+				const data = {
+					email,
+					cashOnHand: 0,
+					positions: [],
+					transactions: [],
+					watchlist: [],
+				};
+				const usersRef = firebase.firestore().collection('users');
+				usersRef
+				.doc(uid)
+				.set(data)
+				.then(() => {
+					navigation.navigate("Login")
+				})					
+			})
+			.catch((error) => setError(error.message))
 	  }
 
     return (
-      <SafeAreaView style={styles.container}>
-				<Image style={styles.image} source={require('../../assets/images/logo.png')} />
-				<View style={{marginBottom: 30}}>
-					<Text>{error}</Text>
-					<View>
-						<Text style={styles.label}>Email Address</Text>
-						<TextInput style={styles.input} autoCorrect={false} autoCapitalize={'none'} onChangeText={(email) => setEmail(email.trim())} />
-					</View>
-					<View>
-						<Text style={styles.label}>Password</Text>
-						<TextInput style={styles.input} secureTextEntry={true} onChangeText={(password) => setPassword(password)} />
-					</View>	
-					<View>
-						<Text style={styles.label}>Confirm Password </Text>
-						<TextInput style={styles.input} secureTextEntry={true} onChangeText={(password) => setConfirmPassword(password)} />
-					</View>				
-					<Text style={styles.forgot} onPress={() => {resetPassword(email)}}>Forgot password or email?</Text>	
-					<Button style={styles.button} onPress={() => onSignUp()} title="Sign Up" />
-				</View>								
-			</SafeAreaView>
+      	<SafeAreaView style={styles.container}>
+			<Image style={styles.image} source={require('../../assets/images/logo.png')} />
+			<View style={{marginBottom: 30}}>
+				<Text>{error}</Text>
+				<View>
+					<Text style={styles.label}>Email Address</Text>
+					<TextInput style={styles.input} autoCorrect={false} autoCapitalize={'none'} onChangeText={(email) => setEmail(email.trim())} />
+				</View>
+				<View>
+					<Text style={styles.label}>Password</Text>
+					<TextInput style={styles.input} secureTextEntry={true} onChangeText={pw => setPassword(pw)} />
+				</View>	
+				<View>
+					<Text style={styles.label}>Confirm Password</Text>
+					<TextInput style={styles.input} secureTextEntry={true} onChangeText={pw => setConfirmPassword(pw)} />
+				</View>				
+				<Text style={styles.forgot} onPress={() => resetPassword(email)}>Forgot password or email?</Text>	
+				<Button style={styles.button} onPress={() => onSignUp()} title="Sign Up" />
+			</View>								
+		</SafeAreaView>
     )
 }
